Rename & clean up db schema builder
